Moved TF files into one directory
