+++ func/tf/CMakeLists.txt
@@ -3,6 +3,7 @@ set(FAASM_USER tf)
 set(TF_DIR /usr/local/code/faasm/third-party/tensorflow/)
 set(TF_LITE_DIR ${TF_DIR}/tensorflow/lite)
 set(TF_LITE_DOWNLOAD_DIR ${TF_LITE_DIR}/tools/make/downloads)
+set(TF_BUILD_DIR ${TF_LITE_DIR}/tools/make/gen/)
 
 # Include TF headers and 3rd party
 include_directories(
@@ -17,18 +18,29 @@ include_directories(
         ${TF_LITE_DOWNLOAD_DIR}/flatbuffers/include
 )
 
+# Set where the library has been built
 if (FAASM_BUILD_TYPE STREQUAL "wasm")
-    function(tf_func exec_name dir_path)
-        faasm_func(${exec_name} ${dir_path})
-        target_link_libraries(${exec_name}
-                /usr/local/code/faasm/third-party/tensorflow/tensorflow/lite/tools/make/gen/wasm32-unknown-none_x86_64/lib/libtensorflow-lite.a)
-    endfunction(tf_func)
+    set(TF_LIB ${TF_BUILD_DIR}/wasm32-unknown-none_x86_64/lib/libtensorflow-lite.a)
 else()
-    function(tf_func exec_name dir_path)
-        faasm_func(${exec_name} ${dir_path})
-        target_link_libraries(${exec_name}
-                /usr/local/code/faasm/third-party/tensorflow/tensorflow/lite/tools/make/gen/linux_x86_64/lib/libtensorflow-lite.a)
-    endfunction(tf_func)
+    set(TF_LIB ${TF_BUILD_DIR}/linux_x86_64/lib/libtensorflow-lite.a)
 endif()
 
-add_subdirectory(image)
+# Build shared library to use with function
+add_library(tf_img_lib
+        bitmap_helpers.cc
+        get_top_n.h
+        get_top_n_impl.h
+        image.h
+        )
+
+target_link_libraries(tf_img_lib ${TF_LIB})
+
+# Add the actual Faasm function
+faasm_func(image image.cc)
+
+# Link with library
+if(FAASM_BUILD_TYPE STREQUAL "knative-native")
+    target_link_libraries(image-knative tf_img_lib)
+else()
+    target_link_libraries(image tf_img_lib)
+endif()
